requirement.txt
fastapi  
uvicorn  
sqlalchemy  
mysql-connector-python  
beautifulsoup4  
requests  